Build environment for a multi-module project
A Makefile for managing the build process is now added.
Most important targets are:

 * make all
 * make build
 * make docker
 * make clean
 * make distclean

When invoking a build target with a `WHAT` argument, a specific module
can be selected.

Example for building virt-launcher only:

   make build WHAT=virt-launcher

`go` and `docker` have their own mechanisms to detect if they need to do
something on command invocation, that means the Makefile does not have
to do the usual tracking of file changes.

Change-Id: Ica69330d80b546876a71176f2089498053a55c84
Signed-off-by: Roman Mohr <rmohr@redhat.com>