Switch to kubernetes client-go Part4
Let the virt-controller handle vm creations only.

It makes sure that we craete pods for every VM definition in the
apiserver. It does this by syncing with the apiserver on startup and
then reacts to the creation of new VM objects.

On successful initialization the virt-controller sets `vm.Status.Phase`
to pending. From this point on virt-handler can take over.

Change-Id: Ibabff637beaa0da1e749692ae112dc62109c34ef
Signed-off-by: Roman Mohr <rmohr@redhat.com>