Scheduling state for virt-controller
virt-controller first creates a pod and then puts the VM into "Scheduling"
state. When the pod got scheduled and is successfully started,
virt-controller moves the VM state then to "Pending". As part of the
last state change it also adds a label "kubevirt.io/nodeName" to the VM,
which can be used for virt-handler to detect VMs it is responsible for.

Change-Id: I7899a894833e084f2a4fe6486350061b63ff7fe5
Signed-off-by: Roman Mohr <rmohr@redhat.com>