Integrate swagger description extraction
When building, the swagger documentation is extracted with the help of
"go:generate". When invoking "make", "go generate" is invoked on all
packages. This command will find all the "go:generate" annotations in
the source files and invoke the commands listed there. For now it is
only used to generate swagger descriptions.

Change-Id: I455acde7e3c85f717fdbf8596cf342bcd142638d
Signed-off-by: Roman Mohr <rmohr@redhat.com>