libvirt: Fake qemu-kube
qemu-kube is responsible for moving a VM into the cgroup of a pod.
But this requires docker and sudo, which are not present in the libvirtd
image. Also access to the hosts docker is needed to query docker.
In addition this will be obsoleted once virt-handler is populating the
cgroup in the domxml.

Considering this it's curently not worth to make the qemu-kube script
work inside the libvirtd container, because it will be obsoleted at
some point.