Assign spice ports and address to VMs and Pods
On the VM side we assign ports starting from 4000 to every spice device.
Futher the spice device has to listen on "0.0.0.0" to make sure that
qemu listens on the Pod IP.
On the Pod side we add a port forward entry for each of the spice
devices, so that we can reach them from outside the Pod.