images: Add a Cockpit Demo pod with KubeVirt provider
This patch introduces a Cockpit pod in _demo shape_.

Due to current limitations, this Cockpit instance is authenticating
against the pod OS, and not against Kubernetes (root:).
Also the wording on the machines page is "host" tuned.

To try this pod:

1. ./cluster/kubectl.sh create -f manifests/cockpit.json
2. Open https://192.168.200.2:9091 (9090 is used by master node cockpit)
3. Navigate to Machines page
4. No Vm should be shown
5. ./cluster/kubectl.sh create -f cluster/vm.json
6. Wait a few seconds and the testvm should appear

The long term solution is to come up with a dashboard for VMs
similar to the Kubernetes dashboard.