Don't export iso as read only
Newer qemu properly checks if the iscsi device is read-only.  We would
prefer not to write to the disk, but fixing that is another topic. For
now, like with the old qemu version which did not to the checks
properly, export the disks read write and write on them, to allow
running the functional tests.