Remove username/password command line args
Providing passwords via command line arguments is a security
flaw because this is visible to any user/service on the host.

Remove the '-user' and '-pass' options from virt-handler
and virt-manager processes. These should not be needed in
general since the libvirt image used with kubevirt is not
going to have authentication turned on for UNIX sockets
anyway.

If someone does need to provide a username/password, then
it is possible todo it out of band via the libvirt client
side auth config file

  http://libvirt.org/auth.html#Auth_client_config

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>