Remove various attributes from the domain interface object
The domain interface must not include any reference to host
devices, so the InterfaceSource is removed entirely. It will
eventually be replaced by something that references a k8s
resource that describes a network connection.

The Alias is an readonly attribute generated by libvirt each
time a guest is started, or device hotplugged. While there
might be a need to use this internally within the system it
should not be exposed to users in the API.

The FilterRef is the name of a libvirt network filter object.
If & when these are needed, those objects will have to be
directly created/deleted based off some k8s resource. As
such the filter reference name should not be part of the
public API

The InterfaceTarget is used to provided the desired name of
the tap device on the physical host. Allowing users to pick
this will cause conflicts between guests as the naming must
be unique globally across the host. The only safe thing is
to rely on dynamically chosen naming.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>