Remove various attributes from the domain OS object
The <boot dev=XXX> syntax is a deprecated way to control the
boot device order. The modern way is to set the boot index
on individual devices. Since we can rely on having modern
QEMU, we should never use the legacy approach for boot ordering.

The domain OS must never include any references to host
file resources, so the NVRam object must be deleted entirely.

The BIOS and Loader objects are deleted because they are
just empty placeholders. They can be re-added if & when
they are needed.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>