Stop moving QEMU into different network namespaces
Moving QEMU into a different network namespaces is not a
supported usage of libvirt and will break in future. For
example when libvirt switches to passing pre-opened for
the listener socket for VNC/SPICE backends, the listening
socket will now be on an IP associated with the libvirtd
POD, despite QEMU being moved into a different network
namespace.

With the network namespace switch removed, the SPICE API
has to be updated to reflect the new IP, and so this info
is obtained from the VM status graphics data added in the
previous commit. The code for changing SPICE IP address
during migration is removed on the assumption that the
SPICE server is set to listen on 0.0.0.0. This is the
implicit default set by the virt-controller.

Since the libvirtd POD is currently using the host network
namespace, it cannot resolve k8s DNS names. Thus when we
configure QEMU to access iSCSI servers, we must resolve
the DNS name to an IP address. To enable this virt-handler
must stay in a POD private network namespace.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>