Remove type attribute from the domain console object
The type of backend used for the domain console object
is not something that should be under control of the
tenant user. Currently the code assumes that the
config is using 'pty', in order for virDomainOpenConsole
to work. In future, it may instead be exposed via a
TCP socket backend. Either way, the code has a specific
backend that will be mandatory.