Add an API to expand a DomainSpec from a libvirt domain XML doc
The DomainSpec provided by the user can be a fairly minimal
representation of a guest. Libvirt will fill in a large number
of attributes, mostly related to the guest ABI, such as device
addresses, disk controllers and other defaults. This data needs
to be propagated back into the DomainSpec. Meanwhile the
DomainSpec contains a bunch of data for which there is no
equivalent in the libvirt XML. For example, disks will have
references to the PersistentVolumeClaims. So it is not possible
to build a DomainSpec from scratch, soley based on libvirt
domainXML. The API thus accepts an input DomainSpec, and then
augments its information with data extracted from the libvirt
XML

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>