Remove various attributes from the domain graphics object
The graphics port and listening address is something that
is to be controlled by the virt handler and/or libvirt
and QEMU directly. Thus should not not be specified in
the public VM domain spec. Likewise it is inteded that
graphics will always be run in secure mode, never plain
text, so again that isn't something that needs to be
exposed as a configuration option.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>