Change source specification for domain disk object
The current domain disk object source directly exposes
the libvirt disk source specification. This is a security
flaw because it allows users to access arbitrary host
files from their guest. This changes the disk source
so that it follows the approach for the VolumeSource
struct in the core k8s API. For now it just supports a
persistent volume claim reference, or a direct ISCSI
server reference.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>