Change MapPersistentVolumes to return an internal domain spec
The MapPersistentVolumes currently accepts a public API VM spec
and returns the same. A lot of fields from the public VM spec,
however, are going to be removed. The only way to set these
fields will be using the internal spec objects. Thus the
MapPersistentVolumes method is changed to return an internal
DomainSpec instead, and renamed to generalize its purpose.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>