docs: update VM config and instance proposal
We update the proposal to highlight few facts (in addition to typo
fixes):

a) A single `VirtualMachineConfig` can be used to create multiple `VM`
instances.
b) A `VM` can be created from `VirtualMachineConfig` or
directly through the API. In case of direct creation, the user must
not manually reference any `VirtualMachineConfig`.

Signed-off-by: Martin Polednik <mpolednik@redhat.com>