Add a design document describing options for managing storage
This document describes the current approach of using PVCs or network block
devices for persistent storage. It then illustrates how this concept can be
extended to enable the notion of ephemeral storage to be supported. Finally
it describes at a very high level, how the PVCs can be populated using a number
of techniques.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>