Delete the virtctl convert-spec command
The convert-spec commands lets users convert the VM Domain
spec between YAML, JSON and XML formats on the client side.
This only works today because the public YAML/JSON schema
and internal XML schemas are identical in structure and
information contained. This is not valid long term because
the public API schema and internal XML are going to diverge.
The public API will contain reference to k8s resources
which will be represented as host files in the XML. This
conversion is lossy and requires integration with k8s
so can't be done standalone in the client.

Rather than continue to provide a command which cannot
deliver what it promises, just delete it.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>