Cloud-Init abstraction with NoCloud data source support
This patch lays the foundation for supporting multiple
cloud-init data sources.

To prove the design, the 'NoCloud' cloud-init data source is
implemented. This data source is the one used by the
fedora atomic project for injecting userdata into atomic
VM instances.

An example of injecting userdata into an ephemeral atomic
VM is provided in the cluster/vm-atomic.yaml VM definition.

Signed-off-by: David Vossel <dvossel@redhat.com>