Automatically bootstrap ginkgo as a part of generate target
Ginkgo tests require each tested package to be bootstrapped to run.
This patch makes the bootstrapping available in makefile's generate
target (which is safe automatically in commit hook).

The nature of this approach causes multiple new files - these indicate
packages that were not really tested by ginkgo. On the other hand,
when adding tests for those packages, we are not required to bootstrap
them anymore.

Signed-off-by: Martin Polednik <mpolednik@redhat.com>