build: Include version and arch in binaries
Previously the binaries were just named according to their name.
To allow users to see what arch and OS they target, this patch is adding
version, OS, and arch to the filename of the cmds.

Before: virtctl
After:  virtctl-v0.0.1-alpha.6-linux-amd64

This filename is used in the cmd/*/* directories only, the names in bin/*
are left untouched to keep a friendly usability.
For compatability a symlink is also added to the cmd/*/* directries to
keep the Dockefiles simple.

The change is mainly intended to keep the version, OS, and arch
informations in case that the binaries get published (like during
releases).

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>