Introduce a temporary way, to retrieve host-specific migration info
Fetch host specific information for migrations from the target host, by
directly querying virt-handler, for the needed information, and provide
it to virt-migrator.

In the future, we either add the required host-specifics during the
migration on the VM directly, or let the target virt-handler start the
migration. If it is better, to collect cluster and host data in an extra
entity, and then start the migration, or let cluster data buble down to
virt-handler, to let it start the migration, is not yet decided.