Prefer building binaries in a docker container
This fixes #436 by creating a docker container for building binaries.
In order for this to fix the aforementioned issue #436, it needs to be
enabled by default.  How is the build done can be selected using
`KUBEVIRT_BUILD_TYPE` environment variable with two possible values:
- `docker` -- build in docker container (default)
- `plain` -- build in current environment

The container is build for this purpose and is left around in order to save
time.  It is cleaned, however, during `make clean`.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>