virtwrap: rename cli package to libvirt package
This is the first commit of a work to provide public API used by other
components to manage virtual machines. At the end it will be prohibed
by the other components to import any packages in virtwrap except
Hypervisor and Guest. The aim is to avoid manipulating low level
details, code redondanced and help maintainability.

- Hypervisor to refer every specific methods related to hypervisor
- Guest to refer every related methods for VMs

Signed-off-by: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@redhat.com>