Add and build kubevirt.yaml
This patch introduces a patch to build a manifest file for the current git tree.
This is especially helpful if we are at a tag. In that case the manifest will create
a manifest which is refencing the tags which are also used for the containers.
Thus these manifests can then be used to deploy this given release on a cluster.

This patch includes a make target to run the functests against the deployment
based on kubevirt.yaml.
This patch depends on #560.

Flow:

```
$ make kubevirt.yaml
$ minikube start ...
$ kubectl create -f kubevirt.yaml
```

and for tagged releases.

```
$ kubectl create -f https://github.com/kubevirt/kubevirt/releases/download/v0.0.4/kubevirt.yaml
```

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>