implement `ephemeral` VolumeSource API
This patch implements bulk of the API used by ephemeral VolumeSources.
We use `ephemeral` as a wrapper of regular VolumeSource, where the
wrapper is responsible for creation and provisioning of the path to
the COW image backed by the nested storage.

Although the paths are setup by this code, we still need to provide
the COW image itself as libvirt's backingStore doesn't do that.

Signed-off-by: Martin Polednik <mpolednik@redhat.com>