api: let user select volume mode
Starting with version 1.9, k8s gained the option to expose the volumes
as raw block device, instead of always creating filesystem on them.
We want to expose this feature for consumption to VMs running into PODs.

This patch implements the minimal API change, adding one optional
parameter to volumes, to turn them in raw block devices.

We should have this feature opt-in because it is alpha-stage in
kubernetes 1.9, enabled using the feature gate. Furthermore,
changes may likely happen during the k8s 1.10 cycle.

Further references:
- https://github.com/kubernetes/community/blob/master/contributors/design-proposals/storage/raw-block-pv.md
- https://schd.ws/hosted_files/kccncna17/8e/Mitsuhiro_Tanino_Block_Volume_KC_CNC_NA17.pdf

Signed-off-by: Francesco Romani <fromani@redhat.com>