Remove defined and stopping event
The defined and stopping event where not directly related to the
matching libvirt event. They were just triggered when specific
conditions were met. IN case of "defined" it got triggered on every
sync.

Further add tests to ensure that all other events, namepley "Started",
"Stopped" and "Crashed" are triggered when we expect it.

Signed-off-by: Roman Mohr <rmohr@redhat.com>