vnc: add timeout to listen
virtctl vnc uses remote-viewer under the hood. If it fails to
spawn remote-viewer for whatever reason, the tool will hang forever
waiting for a connection that will never happen.

A past attempt to fix this made things worse breaking the succesfull
connection flow.
A better fix is done in this patch: we introduce a timeout for
accepting connection to the local socket.

The timeout is set to 60 seconds. Additionally, when verbosiness >= 2,
we also emit how it took to establish the connection. This will
provide a slighter better UX (users can somehow predict how long it
will take to connect after few attempts) and will allow us
to set shorter timeouts in the future.

For reference, on my system it takes ~25s to connect to vnc.