vnc: don't try to cleanup rogue childs
We observed that in some unclear circumstances `virtctl vnc` may
leak remote-viewer processes. In those cases, the remote-viewer
may outlive the virtctl process. Later, remote-viewer will terminate,
most likely for some timeout, so the clutter is temporary.

We had a half-baked fix for this case, but considering that the
scenario is unclear and the fix half-done, better to remove
it completely and to redo it cleanly when the issue is better
understood.