Add networks and interfaces to VM api schema
This patch introduces Networks and Interface entities.
Networks - A list of networks that represent a physical device that can be
attached to the Virtual Machine network interface.
Interfaces - A list of Virtual interfaces that can be created for the virtual
machine. The interface also provides a mechanism that will be used to connected
to the physical device.

Signed-off-by: Vladik Romanovsky <vromanso@redhat.com>