Ensure that filled jinja2 templates are identical to released ones
Very simple test based on diff to ensure that the jinja2 templates stay
in sync with the release manifests based on diff. Because j2cli tends to
"clean" files unasked and it is very easy to write too much newlines in
go templates the input cleaning for "diff" looks a little complex.