Add webhook checks and func test for Disk Serial Number
Set some limits on what we allow for a serial number, libvirt only
considers A-Z,a-z,0-9,_,- so that's easy enough, but we also need
to put an aribtrary limit on the lenght of the string.  Setting a
sane limit to 256 chars.

We added a functional test as well which included some extra functions
in the storage_test module to support the use of Fedora.

Fixes #1133

Signed-off-by: John Griffith <jgriffith@redhat.com>