Fixed VMI phase on shutdown
Before this fix the combination of a non-final vmi and a missing domain
was considered to be unexpected and resulted in a "failed" vmi state,
even when the domain was shut down on purpose.
In order to indicate that a shutdown was triggered, and a missing domain
is a valid thing (it was already deleted), this change introduces a new
vmi phase "stopping", which is set when a shutdown was triggered.
Now we can assume that a vmi without a domain, but in state "stopping",
means that the domain was correctly shut down, and the vmi state can be
set to "succeeded".

Signed-off-by: Marc Sluiter <msluiter@redhat.com>