Let the VM controller only delete VMIs once
If a VMI is already marked for deleteion, the VM controller is not
allowed to delete the VMI again. If it does we can end up in a rythm
where the vmi controller is always behind the latest modification of the
object and can never remove our finalizer once the vmi pod is gone.