Handle delegateIp == false for pod interfaces
Ensure the pod still has external IP connectivity using the IP address
allocated to the pod. This involves moving the allocated IP address to the
bridge (br1), and also restoring the default gateway.

If it's set to false then we don't move MAC / IP addresses to the guest.
Instead, we move the IP address to the br1 bridge to retain connectivity. We
also need to restore the default gateway after setting br1.

Finally, since we don't want to move the allocated MAC address to the guest
either, we don't set it in domain XML.