Add support for Network Disks
This change adds the ability to specify a Network (http) source
as a Volume attachment to a Kubevirt Instance.  The initial use
case being targetted here is the ability to specify an http link
to an ISO image for building/installing VMI's.

Future enhancements may include direct attachment of other network
sources like rbd, iSCSI etc.