Check if non-userfacing errors exceed a specific rate
Let virt-handler check if a specific amount of non-user facing errors
happen during the recheck period. If yes, mark the node as
unschedulable. If the errors do not re-occur the node will be makred as
ready again. This is necessary to react to non-userfacing errors in the
apimachinery which very often indicate partial connection issues for the
informers, while updating the node still works.