Added functional test to validate delegateIp
The test starts a VMI with delegateIp = false, then waits till it boots, brings
eth0 up, and arping's the pod IP address. If Layer2 connectivity is present, we
should receive a reply. We confirm the reply is confirmed by checking the
output of arping call as well as its return code (it's 0 when a reply is
received and 1 when it fails). Finally, the test pings an external service from
inside the pod to validate that connectivity is intact there too.