presets: allow to sort by annotations
Partially addresses https://github.com/kubevirt/kubevirt/issues/1375

We would like to let the operator have control on the order of which
the presets are applied. To do so, we introduce an optional "priority"
annotation, that can be used to enforce the order.

No change is expected, thus the order of application is not specified,
when no annotations are used.

Signed-off-by: Francesco Romani <fromani@redhat.com>