preset: partially allow overrides
This patch introduces overrides among preset, changing the
way conflicts are handled.
Previously, two profiles were conflicting - thus incompatible,
if they overlapped in any way.

Starting with this patch we now allow some overrides where the
resource being changed can be easily compared and when it is
feasible to pick the largest request.

For example, if a preset requires 4 CPU cores and another 2 CPU cores,
we now consider the two presets compatible, and we set the
largest value in the DomainSpec, regardless the order on which
we apply the presets - which is still unspecified.

More portions of the DomainSpec can be made compatible, for
example the CPU model, but this require more logic to do the comparison,
or the help of external libraries (libvirt).
This will be implemented in later patches.

Signed-off-by: Francesco Romani <fromani@redhat.com>