Setup /dev/kvm and /dev/net/tun via virt-handler
In case that the target cointainer does not already contain /dev/kvm or
/dev/net/tun, virt-handler will create the device and whitelist the use
of it for the process inside the container.

In case that /dev/kvm or /dev/net/tun is already there, virt-handler
may only adjust permissions. This allows supporting different external
mechanisms which may provide these devices. Two possible mechanisms:

 * OCI plugins which add the devices to every container
 * k8s device plugins which add it some containers