api: missing V1InterfaceBridge type
After python client generation using current api definition there is
type missing which makes it impossible to manage VMIs. The missing type
seems to be placeholder and it do not contain any definition. It seems
like it is not used and content for it is unknown so changing it to be
a string fixes the issue.


Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>