Handle the (unlikely) event of spoofed MAC address being same
When we spoof a MAC address, we keep the first 3 octets the same (to keep the
vendor) but randomize the last 3 octets. We never guard against the (unlikely,
1 in 256^3 chance) event of the new address to be the same as permanent. This
may result in pod interface and the bridge carrying the same MAC address as the
VMI, which will break traffic flow.

This patch makes us try 10 times if the MAC address is for some reason returned
the same, and gracefully fail if we are so unlucky that we hit the jackpot of
256^(3*10). Perhaps at this point we should also buy a lottery ticket after
returning the error, but I will leave this for a follow-up patch... ;)

Related to bug 1501