Fix double request to the api server
On occasion the double request caused by
calling post and get in fast sequence caused
validating webhook to panic.
The reson for this is due to not enough time
to propage the new object to cache.

Luckily in all cases the second call is
not required for the test to be correct.

The issue is fixed by replacing raw RestClient
by VirtualMachine wrapper and dropping the
second call.

Signed-off-by: Petr Kotas <petr.kotas@gmail.com>