Remove "/kubevirt" from manifest_docker_prefix
It appears that the external provider.sh script is adding an additional
string to the manifest_docker_prefix ("/kubevirt").  The result is that
the image specified in the built manifests is of the form
"$DOCKER_PREFIX/kubevirt/virt-api", however images were pushed to
"$DOCKER_PREFIX/virt-api" so the image pull fails when you try to
deploy.

This change just removes the extra string that was added.