Refactor and extend tests.RunCommand() function
I noticed that some tests are using just stdout from the
tests.RunCommand() function call and use just stdout, while the others
require combined output.  This distinction was solved by having a rather
obscure switch within the function that called cmd.Output() for
"kubectl" command and cmd.CombinedOutput() for the "oc" and "virtctl"
commands.

I have decided to refactor the function in a way that it always return
separate stdout and stderr strings.  This way, if one needs to print
combined output, it is still possible to join the output strings outside
the call:

   out, stderr, err := tests.RunCommand(command, args...)
   if err != nil {
       fmt.Printf("command failed: output %s: %v", out + stderr, err)
   }

In the situation when one needs stdout only, they can simply ignore the
return value:

   out, _, err := tests.RunCommand(command, args...)
   if err != nil {
       fmt.Printf("command failed: output %s: %v", out, err)
   }

This refactoring also allows the third scenario when we would like to
clearly distinguish between stdout and stderr in the consequent error
message:

   out, stderr, err := tests.RunCommand(command, args...)
   if err !- nil {
       fmt.Printf("command failed: output %s: stderr: %s: %v",
           out, stderr, err)
   }

The message that is being passed by function to the "log.Log" calls was
extended and I have also covered all possible errors of the function by
adding two more "log.Log" calls.

External commands are not confined to an explicit namespace, unless "-n"
argument is added.  Since the original version of the tests.RunCommand()
function did not care about namespaces, it was susceptible to produce
errors that are hard to debug, e.g., creating a VMI in the 'default'
namespace instead of 'kubevirt-test-default'.  The extended version of
the tests.RunCommand() function also automatically adds "-n" argument to
confine the command to tests.NamespaceTestDefault namespace.  If one
would like to run external command in different namespace, e.g.,
tests.NamespaceTestAlternative, I have introduced a new function for
this purpose:

  tests.RunCommandWithNS(namespace, command, args...)

If it is not recommendable to place "-n" as the last argument, e.g., in
case of the 'kubectl exec ... -- ...' command, it is possible to
knock-off the automatic "-n" inclusion by passing an empty string:

  tests.RunCommandWithNS("", command, args...)