Add support for running piped external commands
Tests will benefit from a possibility to run external commands in pipe.
For instance, an operator wants to create a VM from Template by a single
step:

$ oc process -f vm-template-fedora.yaml -p NAME=testvm | oc create -f -

rather than a sequence of commands that is implemented now:

$ oc process -f vm-template-fedora.yaml > /tmp/testvm.yaml
$ oc create -f /tmp/testvm.yaml
$ rm /tmp/testvm.yaml

The newly added tests.RunCommandPipe() function takes any number of
string arrays that represents arg arrays of the individual commands
which are supposed to be coupled into stdout pipe chain.  The 0th
element of each arg array is interpreted as command name.  The function
returns error when it receives less than two arguments.

The stderr output of all commands is merged into single buffer.
Function returns the output of the last command in chain (converted to
string), co-joined stderr buffer (converted to string), and error.

This newly added function also adds "-n" argument to all commands
executed in pipe, confining them to the tests.NamespaceTestDefault
namespace.  For running command pipe in different namespace one can
call:

  tests.RunCommandPipeWithNS(namespace, commands...)

If the command chain contains a command for which it is not
recommendable to place "-n" as the last argument, it is possible to
knock-off the automatic namespace inclusion by passing an empty string:

  tests.RunCommandPipeWithNS("", commands...)

Note that in such case the namespace has to be passed via argument
to *each command* separately.