Add collectd node agent to export VMI metrics
In order to monitor, troubleshoot and tune VMs, we would like
to expose detailed metrics about how the VM instance is behaving.
We are concerned about the infrastructure here, so the infrastructure
processes running inside the virt-launcher pod.
For example, we want to assess the overhead of the aformenentioned
infrastructure, allowing to have answers to queries like

- how much memory is libvirtd taking?
- is that memory growing?
- how much memory QEMU takes?
- is libvirtd leaking threads?

This patch provide the initial support to report those metrics.

Signed-off-by: Francesco Romani <fromani@redhat.com>