Don't deploy test infra if using external provider
This patch modifies cluster/deploy.sh to NOT attempt deploying test
infra when using an external KubeVirt Provider.

Since we don't have control over the target Cluster in this case there's
a high probability that it won't have all of the expected configuration
settings and avaialble resources we expect for our tests.  Rather than
just running it and having it fail for various reasons, this patch will
skip that portion of the deployment for external providers.

In the case of a user that knows what they're doing and has everything
he/she needs on their cluster, they can override this behavior with an
enviornment variable.

`KUBEVIRT_DEPLOY_TESTS_TO_EXTERNAL_PROVIDER=true` will override the
provider check and attempt to deploy the additional test infra.