test.sh: replace locking mechanism
Instead of using a spin lock, which might leave the lock behind if the
process was killed with SIGKILL, we will use flock which releases the
process as soon as the process dies.

It's needed now because we're planning to abort tests on a specific PR
when the PR is being updated, the the abort is not guaranteed to be
graceful.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>