Updated networkName description in API schema
The comment incorrectly claims that the network is always in the
user's namespace. In fact, we do support networks in different
namespaces. As we should, because de-facto network CRD standard
document allows to attach pods to networks from other namespaces. We
also have a functional test that makes sure it works.