prometheus: virt-handler: expose vm metrics
This patch adds a prometheus Collector
(https://godoc.org/github.com/prometheus/client_golang/prometheus#Collector)
to fetch and expose the VM metrics.
Internally, we make use of the GetDomainStats cmd-client command
to gether the metrics.

Unlike the other prometheus metrics, we need to do an explicit
initialization step, because we need to pass around the virtShareDir.

Signed-off-by: Francesco Romani <fromani@redhat.com>