Improve preStart hook by taking arbitrary xml modifications
Modifications are not unmarshalled and directly sent to libvirt after
being processed by the webstart hook.

Since e.g. the metadata can under some conditions change, and we persist
the whole domain xml with the change, we need to call the prestart hook
also in these scenarios.

The prestart hooks needed to be written in an idempotent way because
they are executed at-least-once already. It should therefore not
introduce problems with existing prestart hooks. In the future we should
not write kubevirt specific metadata. We can directly persist it in
memory of virt-launcher.