Make compute container 1st in the list for launcher pods
The new multus mutating webhook automatically registers requests /
limits for devices allocated by device plugins based on network
attachment CRD resourceName annotation.

To accommodate for the webhook, this patch shuffles the order of
launcher pod containers accordingly.

I believe it's safe because I am not aware of any other component that
would assume the order of containers in the launcher pod that would
rely on it.