Request resources listed in multus network resourceName annotation
Whenever a network has the 'resourceName' annotation, it means that
the listed resource implements backend for the network. In this case,
we would like to automatically request the listed resource for
launcher container.

In the future, Multus mutating webhook will handle the job. This is
implemented in https://github.com/intel/multus-cni/pull/186 up for
review. That being said, the multus webhook will probably need some
rework. Specifically, there are plans to split it into two pieces, one
validating and another mutating, and contribute them to separate
repos. This may take quite some time, so in the meantime, we are going
to introduce this kubevirt specific workaround. Once multus team
releases an official mutating hook, we will adopt it in kubevirt and
remove the workaround.