Refactor resolv.conf via cloudinit.
Adding dns via network-config only seems to be supported for sysconfig
style OSs. It also only appends certain elements to resolv.conf instead
of overwriting it. This change appends the resolv_conf yaml to userData
if it exists and is of type #config.

Add domain parsing to the util/net/dns package.

Signed-off-by: Greg Bock <greg.bock@stackpath.com>